+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 
 - Fixed bug where an opened collapsed tab would close when clicking the container without clicking a list item.
 - Fixed issue with disabled loading button not displaying properly (#61).
+- Fixed a bug that would cause the rangeslider script to generate one style tag for each rangeslider component on a given page.
 - Fixed an issue with faulty grid baserules.
 
 ### Removed